+++ NotesApp/Views/SavedDocumentsView.swift
@@ -7,24 +7,6 @@
 
 import SwiftUI
 
-struct Document: Codable, Identifiable {
-    var id = UUID()
-    var name: String = "Document name"
-    var text: String = "# Welcome"
-    var lastModified = Date.now
-
-    init(name: String, text: String) {
-        self.name = name
-        self.text = text
-    }
-
-    static func == (lhs: Document, rhs: Document) -> Bool {
-        lhs.id      == rhs.id &&
-        lhs.name    == rhs.name &&
-        lhs.text    == lhs.text
-    }
-}
-
 struct SavedDocumentsView: View {
 
     @State var columnCount: Int = 3
